+++ index.js
@@ -1,11 +1,7 @@
 module.exports = {
 	nodes: [
 		require('./dist/nodes/Notifuse/Notifuse.node.js'),
-		require('./dist/nodes/NotifuseContactTrigger/NotifuseContactTrigger.node.js'),
-		require('./dist/nodes/NotifuseListTrigger/NotifuseListTrigger.node.js'),
-		require('./dist/nodes/NotifuseSegmentTrigger/NotifuseSegmentTrigger.node.js'),
-		require('./dist/nodes/NotifuseEmailTrigger/NotifuseEmailTrigger.node.js'),
-		require('./dist/nodes/NotifuseCustomEventTrigger/NotifuseCustomEventTrigger.node.js'),
+		require('./dist/nodes/NotifuseTrigger/NotifuseTrigger.node.js'),
 	],
 	credentials: [require('./dist/credentials/NotifuseApi.credentials.js')],
 };